Day10/Exercise1/requirements.txt
kafka-python==2.0.2
confluent-kafka==2.1.1
pyspark==3.4.1
findspark==2.0.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
pandas==2.0.3
numpy==1.24.4
sqlalchemy==2.0.20
kestra-client==0.1.2 